revert int64_t changes

+++ torch/csrc/jit/codegen/cuda/arith.cpp
@@ -395,7 +395,7 @@ static TensorView* newForReduction(
 
 TensorView* reductionOp(
     BinaryOpType reduction_op_type,
-    const std::vector<int64_t>& axes,
+    const std::vector<int>& axes,
     Val* init,
     TensorView* tv) {
   TORCH_CHECK(
@@ -407,9 +407,9 @@ TensorView* reductionOp(
       "Reducing a tensor once it's gone under transformations is not permitted at this time. Please set reductions before calling split/merge/computeAt.");
 
   std::vector<unsigned int> uint_axes;
-  for (int64_t axis : axes) {
+  for (int axis : axes) {
     if (axis < 0)
-      axis += int64_t(tv->nDims());
+      axis += int(tv->nDims());
 
     TORCH_CHECK(
         axis >= 0 && (unsigned int)axis < tv->nDims(),
@@ -429,9 +429,7 @@ TensorView* reductionOp(
   return out;
 }
 
-TORCH_CUDA_API TensorView* sum(
-    TensorView* v1,
-    const std::vector<int64_t>& axes) {
+TORCH_CUDA_API TensorView* sum(TensorView* v1, const std::vector<int>& axes) {
   Val* init;
   switch (v1->getDataType().value()) {
     case (DataType::Float):
